[JENKINS-16364] Update support for Build Dsl job type; refactor naming conventions.
Note that the dsl block must be triple quoted if it's pretty printed.

Updated usage below.

Usage:

  buildDslBlock( <build dsl text, escaped or flat> )

Example (simple):

  buildDslBlock('build ("job1")')

Example (pretty):

  buildDslBlock("""
    parallel(
        { " job1 " }
    )
    """)

Example (passing in a variable's value, so the created job has
the value in the text of the dsl block):

  buildDslBlock('build("${BUILD_TAG}")')
  End result in build DSL node: build("my-base-job-dsl-job-12")

Example (passing in a variable, so the end-result is a created job
that has the variable (and not its value)):

  buildDslBlock('build("\${BUILD_TAG}")')
  End result in build DSL node: build("${BUILD_TAG}")