started to port DslScriptLoader back to Java
The Groovy version causes a regression when executing script files with a
name that is not a valid Java class name, see GROOVY-4020. E.g. when
running a script called jobs-blog.groovy the following exception is thrown:

java.lang.ClassFormatError:
Illegal class name "jobs-blog$run" in class file jobs-blog$run

One would assume that this should also be a problem in the Java version of
DslScriptLoader, but for some reason it isn't.