Fixes use of jobManagement inside a conditionalStepsContext
In 8bfeba6, cleaning up of context constructors hid the visibility of
jobManagmenet. In practice this is a good idea to limit the visibility
of private contexts.

This unintentionally breaks JobDsl extension when used inside a
conditionalSteps (with Categories or metaClass patching) due to the fact
that a conditionalStep no longer has access to the jobManagement field.