Add support for later versions of config-file-provider-plugin.
More recent version (starting with 2.8.1) have a different constructor
signature for MavenSettingsConfig.

Switched to using a data binder similar to StaplerRequest.bind to
support any constructor signature as long as it is annotated with
@DataBoundConstructor.

Constructor parameters are now provided as a Map and all supported
parameters will be passed to the constructor.