When ghprb version is new enough, emit configVersion 3
Earlier versions of the ghprb did not have regex matching for trigger phrases.  When this was introduced, the config upgrade process quoted the trigger phrase strings.  Because the DSL does not emit configVersion 3, a regex trigger phrase will get quoted when it doesn't need to even to.  To preserve backwards compatible behavior, a new element 'regexTriggerPhrase' is introduced.  When regexTriggerPhrase is used, the ouptut triggerPhrase is not quoted.  When the original triggerPhrase is used, it is automatically wrapped in \Q \E

Emit configVersion 3 if plugin is new enough, but quote trigger phrases if triggerPhrase node is used.  If regexTriggerPhrase is used, do not quote

Fixup spacing