JENKINS-41308 add possibility to set sandbox in activeChoiceParam script
The ActiveChoiceParameter plugin already allows the setting of `scripts` and `fallbackScripts` to be run in a `sandbox` via the UI. With this commit, the ability to set sandboxing via the Job-DSL is introduced.

If the `sandbox` property is not present, it defaults to the old mechanism. In case an empty closure is added to either `script` or `fallbackScript`, sandboxing is also disabled. Only if `sandbox` is explicetly set to `true` or if the `sandbox` property is set without explicitly mentioning the state (e.g. `sandbox()`), sandboxing is enabled.

All previous tests and ways of functioning are unaffected by this chagne.