Fix: Don't destroy credentials when re-creating folders
As noted in JENKINS-44681, any credentials on folders are lost when
re-creating folders, as we don't merge the existing credentials with the
newly created folder.

To do this, we need to find any instances of
`FolderCredentialsProvider.FolderCredentialsProperty` in the `Folder`'s
and if present, merge the credentials.

We need to promote our dependency on `cloudbees-folder` to a runtime
dependency, as we now need to reference the `Folder` and any
`FolderCredentialsProperty`s.

Note that in `JenkinsJobManagementSpec`, we need to use
`getNodeTemplate` to ensure that the real implementation of `getXml` is
called, which will execute the `configuredBlocks`.

Because we need to call to the `configure` method, we need to call out
to a new Groovy file, `DslItemConfigurer`, which processes the new
`AbstractFolderProperty` and converts it to the correct XML
representation. When parsing it, we need to make sure we add a valid XML
header, otherwise `XmlParser` will reject it.

Closes JENKINS-44681.