Fix: Copy the API Viewer resources into the built plugin
As mentioned in JENKINS-66988, we currently are not producing the API
Viewer in the built plugin.

This is due to [0], in which the JPI plugin ignores any processing of
the WAR that we do separately.

As an intermediate solution, we can make sure that we copy the files
into the `src/main/webapp` directory, so it is present for the plugin.

This requires that we then exclude the files from Git, as we want them
to only be generated, but we need to make sure that `config.json` is
still included.

Finally, we make sure that the `jpi` task depends on the copying of
files, for standalone building of the plugin, and the `copyApiViewer`
depends on the build process for the `api-viewer` project.

[0]: https://github.com/jenkinsci/gradle-jpi-plugin/issues/193