Fix: Don't destroy credentials when re-creating organization folders
Similar to 8d8ea63fcefe0028c5168dffd9210955c41fe773, we should be able
to support organization folders not losing their credentials.

This requires we bump our dependency of `credentials` to pick up the
`branch-api` dependency.

We need to handle the various bits of metadata available in the
`OrganizationFolder`s, so our tests are slightly more complex than we
usually have.

Co-authored-by: Jean-François Roche <jfroche@affinitic.be>