Use the Jenkins plugin BOM
To better manage dependencies of various libraries, we can use the
Parent POM for Jenkins Plugins, which we can import as a Platform, and
then manage our dependencies more appropriately.

We can use 3.57 as the closest plugin version that maps to our version
of Jenkins core.

As a step towards JENKINS-68275.