Add a job to sync docs with GitHub Wiki
Right now, it's a manual step I perform when I'm doing a release, which
leads to this being out-of-sync.

We can fortunately automate it following comments in [0] to do this more
easily.

We want it to allow to fail, in the cases that we're in sync, which is
most likely.

And we want to make sure that this only executes on the main branch.

[0]: https://github.com/orgs/community/discussions/25929