Replace cg-lib based mocks with byte-buddy
cg-lib is a dead project that does not support recent JDK versions.

This switches to create stubs in spock with byte-buddy.  The PR does not
directly depend on byte-buddy but rather obtains it transitively via the
managed dependency of mockito in the plugin-pom.

This allows it to be kept more up-to-data and in a compatable version
(as it also brings in the Objenisis version at a know good version).