Fix unit tests in latest BOM builds (#1542)
Replace cglib-based mocks with Byte Buddy

cglib is a dead project that does not support recent JDK versions.

This switches to create stubs in Spock with Byte Buddy. The PR does not
directly depend on Byte Buddy but rather obtains it transitively via the
managed dependency of Mockito in the plugin parent POM.

This allows it to be kept more up-to-date and in a compatible version
(as it also brings in the Objenesis version at a known good version).